requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.26.5
importlib-resources==6.1.1
inflection==0.5.1
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
pkgutil-resolve-name==1.3.10
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
rpds-py==0.12.0
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
zipp==3.17.0
